Moved convertToMultiType method to GmlUtils and added JUnit tests for same.
