+++ src/test/resources/gml/SurfaceGeometry.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<gml:geometryMember xmlns:gml="http://www.opengis.net/gml/3.2">
+  <gml:Surface gml:id="Surface_NL.KAD.CP.LNK00D.2377.1" srsName="urn:ogc:def:crs:EPSG::4258">
+    <gml:patches>
+      <gml:PolygonPatch>
+        <gml:exterior>
+          <gml:LinearRing>
+            <gml:posList>
+            52.266657 6.932919 52.266749 6.931853 52.267056 6.928073 52.267144 6.926986
+            52.267412 6.927115 52.267771 6.927283 52.268067 6.927014 52.268291 6.926723
+            52.268542 6.926403 52.270135 6.927449 52.270171 6.927303 52.270260 6.927343
+            52.270862 6.927612 52.271038 6.927691 52.271116 6.927470 52.271214 6.927192
+            52.271319 6.927262 52.271635 6.927474 52.271585 6.928170 52.271524 6.928807
+            52.271514 6.928913 52.271482 6.929246 52.271355 6.930484 52.271355 6.930499
+            52.271487 6.930543 52.272250 6.930778 52.272739 6.930936 52.273093 6.931046
+            52.273315 6.931114 52.273825 6.931278 52.273881 6.931291 52.273704 6.932439
+            52.273553 6.933403 52.273411 6.934301 52.271825 6.933974 52.270937 6.933783
+            52.270107 6.933611 52.269130 6.933415 52.268728 6.933338 52.267354 6.933061
+            52.267292 6.933048 52.266657 6.932919
+            </gml:posList>
+          </gml:LinearRing>
+        </gml:exterior>
+      </gml:PolygonPatch>
+    </gml:patches>
+  </gml:Surface>
+</gml:geometryMember>
+
